jpa dao test passed for GiftCertificateDAOImplTest

+++ src/main/java/com/epam/esm/module2boot/dao/jpaImpl/GiftCertificateDaoImpl.java
@@ -4,12 +4,14 @@ import com.epam.esm.module2boot.dao.GiftCertificateDAO;
 import com.epam.esm.module2boot.exception.NotFoundException;
 import com.epam.esm.module2boot.model.GiftCertificate;
 import jakarta.persistence.EntityManager;
+import jakarta.persistence.TypedQuery;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 @Repository
 @Transactional
@@ -32,8 +34,9 @@ public class GiftCertificateDaoImpl implements GiftCertificateDAO {
     }
 
     @Override
-    public boolean deleteGiftCert(int id) {
+    public boolean deleteGiftCert(int id) throws NotFoundException {
         GiftCertificate giftCertificate = entityManager.find(GiftCertificate.class, id);
+        if (giftCertificate == null) throw new NotFoundException("GiftCertificate not found with id:" + id);
         entityManager.remove(giftCertificate);
         entityManager.flush();
         return true;
@@ -54,6 +57,24 @@ public class GiftCertificateDaoImpl implements GiftCertificateDAO {
 
     @Override
     public List<GiftCertificate> getAllByParam(Map<String, Object> params, List<String> sorting) {
-        return null;
+        String whereStr = HQLHelper.getWhereStr(params);
+        String sortStr = HQLHelper.getSortingSubStr(sorting);
+
+        System.out.println("\nqueryStr:" + whereStr);
+        System.out.println("sortStr:" + sortStr + "\n");
+
+        TypedQuery<GiftCertificate> query = entityManager.createQuery(
+                """
+                              SELECT gc FROM GiftCertificate gc
+                              OUTER JOIN gc.tags t
+                        """ + whereStr + sortStr
+                , GiftCertificate.class);
+
+        params.forEach((s, o) ->
+                query.setParameter(HQLHelper.translateParameter(s), o));
+
+        return query.getResultList();
     }
+
+
 }
